Resolve comment apply defer()
