+++ src/app/article-detail/ui/comment-list/comment-list.component.html
@@ -1,25 +1,19 @@
-@for (item of (this.comments); track
-$index) {
-<div class="card">
-  <div class="card-block">
-    <p class="card-text">
-      {{ item.body }}
-    </p>
+@for (item of this.comments; track item.id) {
+  <div class="card">
+    <div class="card-block">
+      <p class="card-text">
+        {{ item.body }}
+      </p>
+    </div>
+    <div class="card-footer">
+      <a routerLink="/profile/author" class="comment-author">
+        <img src="http://i.imgur.com/Qr71crq.jpg" class="comment-author-img" />
+      </a>
+      &nbsp;
+      <a routerLink="/profile/jacob-schmidt" class="comment-author">{{
+        item.author.username
+      }}</a>
+      <span class="date-posted">{{ item.createdAt | date: "mediumDate" }}</span>
+    </div>
   </div>
-  <div class="card-footer">
-    <a href="/profile/author" class="comment-author">
-      <img
-        src="http://i.imgur.com/Qr71crq.jpg"
-        class="comment-author-img"
-      />
-    </a>
-    &nbsp;
-    <a href="/profile/jacob-schmidt" class="comment-author">{{
-      item.author.username
-    }}</a>
-    <span class="date-posted">{{
-      item.createdAt | date : "mediumDate"
-    }}</span>
-  </div>
-</div>
 }
