+++ src/app/shared/interceptors/api-auth.interceptor.ts
@@ -1,14 +1,11 @@
 import {
-  HttpErrorResponse,
   HttpHandlerFn,
   HttpInterceptorFn,
   HttpRequest,
 } from '@angular/common/http';
 import { inject } from '@angular/core';
 import { LocalStorage, STORAGE_KEY } from '../constants';
-import { AuthResponse, ErrorResponse } from '../models';
-import { catchError, EMPTY, throwError } from 'rxjs';
-import { AuthStore } from '../store';
+import { AuthResponse } from '../models';
 
 export const apiAuthInterceptor: HttpInterceptorFn = (
   req: HttpRequest<unknown>,
@@ -31,19 +28,3 @@ export const apiAuthInterceptor: HttpInterceptorFn = (
   return next(req);
 };
 
-export const handleErrorInterceptor: HttpInterceptorFn = (
-  req: HttpRequest<unknown>,
-  next: HttpHandlerFn,
-) => {
-  const authStore = inject(AuthStore);
-  return next(req).pipe(
-    catchError((error: HttpErrorResponse) => {
-      const err = error.error as ErrorResponse;
-      if(err.code === 401 || err.code === 403){
-        authStore.logout();
-      }
-      return throwError(() => error);
-    })
-  )
-
-};
