+++ src/app/shared/interceptors/index.ts
@@ -1,2 +1,3 @@
 export * from './api-auth.interceptor';
 export * from './api-prefix.interceptor';
+export * from './api-handle.interceptor';
